Add the position value and first test
The position value designates the location where the questionnaire field will be. The position is useful to maintain the same order before and after the conversion from csv to json.

The first test ensures that the conversion is carried out correctly. We compare the new json file with the json file that the conversion should make.